+++ src/Button/Button.tsx
@@ -8,8 +8,6 @@ import React from 'react';
 
 import classNames from '../utils/classNames';
 
-import styles from './button.module.scss';
-
 export type ButtonVariants =
     | 'primary'
     | 'secondary'
@@ -17,7 +15,8 @@ export type ButtonVariants =
     | 'info'
     | 'warning'
     | 'danger'
-    | 'link';
+    | 'link'
+    | 'link-button';
 
 type ButtonProps = {
     id?: string;
@@ -39,21 +38,39 @@ const Button: React.FC<ButtonProps> = ({
     title,
     large = false,
 }) => (
-    <button
-        type="button"
-        id={id}
-        className={classNames(
-            styles.button,
-            large && styles.large,
-            styles[variant],
-            className
-        )}
-        disabled={disabled}
-        onClick={onClick}
-        title={title}
-    >
-        {children}
-    </button>
+    <div className="tw-preflight">
+        <button
+            type="button"
+            id={id}
+            className={`${classNames(
+                large
+                    ? 'tw-h-8 tw-px-4 tw-text-sm'
+                    : 'tw-h-6 tw-px-2 tw-text-xs',
+                variant === 'primary' &&
+                    'tw-bg-nordicBlue tw-text-white active:enabled:tw-bg-nordicBlue-700',
+                variant === 'secondary' &&
+                    'tw-border tw-border-gray-700 tw-bg-white tw-text-gray-700  active:enabled:tw-bg-gray-50',
+                variant === 'success' &&
+                    'tw-bg-green tw-text-white  active:enabled:tw-bg-green-700',
+                variant === 'info' &&
+                    'tw-bg-nordicBlue tw-text-white active:enabled:tw-bg-nordicBlue-700',
+                variant === 'warning' &&
+                    'tw-bg-orange tw-text-white active:enabled:tw-bg-orange-700',
+                variant === 'danger' &&
+                    'tw-bg-red tw-text-white active:enabled:tw-bg-red-700',
+                variant === 'link' &&
+                    'tw-bg-transparent tw-p-0 tw-text-nordicBlue',
+                variant === 'link-button' &&
+                    'tw-border tw-border-nordicBlue tw-bg-white tw-text-nordicBlue active:enabled:tw-bg-gray-50',
+                className
+            )}`}
+            disabled={disabled}
+            onClick={onClick}
+            title={title}
+        >
+            {children}
+        </button>
+    </div>
 );
 
 export default Button;